refactor(atoms): merge SfIcon
